feat: rewrite how font options are applied
From now on, options will apply only if their default value is changed.
For example, you can increase font-weight on all websites without having
to specify a custom base font.  It's a small touch which should increase
the usability of this userstyle for a wider range of use-cases.

In addition, the code won't be generated for options that are using
default values.  This way we only inject what we need.  It's a small
optimization that doesn't have a lot of impact, but I thought it would
be a "fun fact" as it highlights an interesting feature of Stylus-lang.